# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Correspondence Workspace API",
    "description": "Built-in mailbox, event-driven case automation, connected system workspaces and human review.",
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/scenarios": {
      "get": {
        "tags": [
          "Synthetic data and references"
        ],
        "summary": "Scenarios",
        "operationId": "scenarios_api_scenarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioRead"
                  },
                  "type": "array",
                  "title": "Response Scenarios Api Scenarios Get"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{scenario_id}/instances": {
      "post": {
        "tags": [
          "Synthetic data and references"
        ],
        "summary": "Load Scenario",
        "operationId": "load_scenario_api_scenarios__scenario_id__instances_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "DEMO-01",
                "DEMO-02",
                "DEMO-03",
                "DEMO-04",
                "DEMO-05"
              ],
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioLoad"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/knowledge": {
      "get": {
        "tags": [
          "Synthetic data and references"
        ],
        "summary": "Knowledge",
        "operationId": "knowledge_api_knowledge_get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "DEMO-01",
                "DEMO-02",
                "DEMO-03",
                "DEMO-04",
                "DEMO-05"
              ],
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "client_code",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "DEMO-NORTH",
                "DEMO-HARBOR"
              ],
              "type": "string",
              "title": "Client Code"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "guidance",
                    "template",
                    "disclosure"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeRead"
                  },
                  "title": "Response Knowledge Api Knowledge Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/taxonomy": {
      "get": {
        "tags": [
          "Synthetic data and references"
        ],
        "summary": "Taxonomy",
        "operationId": "taxonomy_api_taxonomy_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "work_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Work Type"
            }
          },
          {
            "name": "class_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class Name"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxonomyPage"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/reference/summary": {
      "get": {
        "tags": [
          "Synthetic data and references"
        ],
        "summary": "Reference Summary",
        "operationId": "reference_summary_api_reference_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceSummary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/clients": {
      "get": {
        "tags": [
          "Synthetic data and references"
        ],
        "summary": "Clients",
        "operationId": "clients_api_clients_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Clients Api Clients Get"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/tasks": {
      "get": {
        "tags": [
          "Synthetic data and references"
        ],
        "summary": "Tasks",
        "operationId": "tasks_api_cases__case_id__tasks_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "title": "Response Tasks Api Cases  Case Id  Tasks Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cases/{case_id}/concerns": {
      "get": {
        "tags": [
          "Synthetic data and references"
        ],
        "summary": "Concerns",
        "operationId": "concerns_api_cases__case_id__concerns_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConcernRead"
                  },
                  "title": "Response Concerns Api Cases  Case Id  Concerns Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cases/{case_id}/evidence/{evidence_id}/file": {
      "get": {
        "tags": [
          "Synthetic data and references"
        ],
        "summary": "Evidence File",
        "operationId": "evidence_file_api_cases__case_id__evidence__evidence_id__file_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          },
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "simulation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Simulation Id"
          },
          "loan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Loan Id"
          },
          "loan_identifier": {
            "type": "string",
            "title": "Loan Identifier"
          },
          "ccid": {
            "type": "string",
            "title": "Ccid"
          },
          "client_code": {
            "type": "string",
            "title": "Client Code"
          },
          "borrower_display_name": {
            "type": "string",
            "title": "Borrower Display Name"
          },
          "balance_minor": {
            "type": "integer",
            "title": "Balance Minor"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "correspondence_text": {
            "type": "string",
            "title": "Correspondence Text"
          },
          "family": {
            "$ref": "#/components/schemas/CaseFamily"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "original_received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Original Received At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "classification": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification"
          },
          "content_revision": {
            "type": "integer",
            "title": "Content Revision"
          },
          "routing": {
            "additionalProperties": true,
            "type": "object",
            "title": "Routing"
          },
          "pending_work": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pending Work"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "simulation_id",
          "loan_id",
          "loan_identifier",
          "ccid",
          "client_code",
          "borrower_display_name",
          "balance_minor",
          "currency",
          "subject",
          "correspondence_text",
          "family",
          "owner",
          "status",
          "original_received_at",
          "created_at",
          "updated_at",
          "revision",
          "classification",
          "content_revision",
          "routing",
          "pending_work"
        ],
        "title": "CaseRead"
      },
...
      "ConcernRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "case_id": {
            "type": "string",
            "format": "uuid",
            "title": "Case Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "disposition": {
            "type": "string",
            "title": "Disposition"
          },
          "evidence_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Evidence Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "case_id",
          "description",
          "disposition",
          "evidence_ids"
        ],
        "title": "ConcernRead"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "ReferenceSummary": {
        "properties": {
          "taxonomy_count": {
            "type": "integer",
            "title": "Taxonomy Count"
          },
          "knowledge_row_count": {
            "type": "integer",
            "title": "Knowledge Row Count"
          },
          "statuses": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Statuses"
          },
          "curated_item_count": {
            "type": "integer",
            "title": "Curated Item Count"
          },
          "raw_response_text_exported": {
            "type": "boolean",
            "const": false,
            "title": "Raw Response Text Exported"
          },
          "summary_reported_rows": {
            "type": "integer",
            "title": "Summary Reported Rows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "taxonomy_count",
          "knowledge_row_count",
          "statuses",
          "curated_item_count",
          "raw_response_text_exported",
          "summary_reported_rows"
        ],
        "title": "ReferenceSummary"
      },
...
      "ScenarioLoad": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "variant": {
            "type": "string",
            "enum": [
              "base",
              "missing_document",
              "unreadable_document",
              "wrong_loan",
              "followup"
            ],
            "title": "Variant",
            "default": "base"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ScenarioLoad"
      },
      "ScenarioRead": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "enum": [
              "DEMO-01",
              "DEMO-02",
              "DEMO-03",
              "DEMO-04",
              "DEMO-05"
            ],
            "title": "Scenario Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "family": {
            "type": "string",
            "title": "Family"
          },
          "client_code": {
            "type": "string",
            "title": "Client Code"
          },
          "variants": {
            "items": {
              "type": "string",
              "enum": [
                "base",
                "missing_document",
                "unreadable_document",
                "wrong_loan",
                "followup"
              ]
            },
            "type": "array",
            "title": "Variants"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenario_id",
          "version",
          "title",
          "family",
          "client_code",
          "variants"
        ],
        "title": "ScenarioRead"
      },
...
      "TaxonomyPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaxonomyRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "TaxonomyPage"
      },
...
    }
  }
}